+++ inc/opendb.php
@@ -17,6 +17,8 @@ if(strtoupper(getenv('PLATFORM')) === "CLOUD"){
 	$tmp_path = getenv('TMP_PATH') ?: "/tmp/";
 	$allowed_ext = getenv('FILE_EXT') ?: "pdf, doc, docx, xls, csv, xlsx, png, jpg, jpeg, gif";
 	$max_filesize = getenv('MAX_FILE_SIZE') ?: "4194304";
+	$bkparams =  getenv('DB_BACKUP_PARAMETERS') ?: '--column-statistics=1';
+
 
 }elseif(file_exists(__ROOT__.'/inc/config.php') == TRUE) {
 	require_once(__ROOT__.'/inc/config.php');
@@ -25,7 +27,7 @@ if(strtoupper(getenv('PLATFORM')) === "CLOUD"){
 
 
 
-$conn = mysqli_connect($dbhost, $dbuser, $dbpass, $dbname) or die ('Unable to connect to '.$dbname.' database on '.$dbhost.' host. Please make sure the dabase exists and you user has full permissions on it.');
+$conn = mysqli_connect($dbhost, $dbuser, $dbpass, $dbname) or die ('Unable to connect to '.$dbname.' database on '.$dbhost.' host. Please make sure the dabase exists and user '.$dbuser.' has full permissions on it.');
 mysqli_select_db($conn, $dbname);
 mysqli_set_charset($conn, "utf8");