+++ pages/formula.php
@@ -64,16 +64,16 @@ if($form[0]['ingredient']){
             </div>
         <!-- Nav tabs -->
         <ul class="nav nav-tabs" role="tablist">
-          <li class="active"><a href="#main_formula" id="formula_tab" role="tab" data-toggle="tab"><i class="fa fa-bong mr2"></i>Formula</a></li>
-    	  <li><a href="#impact" id="impact_tab" role="tab" data-toggle="tab"><i class="fa fa-magic mr2"></i>Notes Impact</a></li>
-          <li><a href="#pyramid" id="pyramid_tab" role="tab" data-toggle="tab"><i class="fa fa-table mr2"></i>Olfactory Pyramid</a></li>
-          <li><a href="#summary" id="summary_tab" role="tab" data-toggle="tab"><i class="fa fa-cubes mr2"></i>Notes Summary</a></li>
-          <li><a href="#ingRep" id="reps_tab" role="tab" data-toggle="tab"><i class="fa fa-exchange-alt mr2"></i>Replacements</a></li>
-          <li><a href="#attachments" id="attachments_tab" role="tab" data-toggle="tab"><i class="fa fa-paperclip mr2"></i>Attachments</a></li>
-          <li><a href="#revisions" id="revisions_tab" role="tab" data-toggle="tab"><i class="fa fa-clock-rotate-left mr2"></i>Revisions</a></li>
-          <li><a href="#timeline" id="timeline_tab" role="tab" data-toggle="tab"><i class="fa fa-timeline mr2"></i>History</a></li>
+          <li class="active"><a href="#main_formula" id="formula_tab" role="tab" data-toggle="tab"><i class="fa fa-bong mx-2"></i>Formula</a></li>
+    	  <li><a href="#impact" id="impact_tab" role="tab" data-toggle="tab"><i class="fa fa-magic mx-2"></i>Notes Impact</a></li>
+          <li><a href="#pyramid" id="pyramid_tab" role="tab" data-toggle="tab"><i class="fa fa-table mx-2"></i>Olfactory Pyramid</a></li>
+          <li><a href="#summary" id="summary_tab" role="tab" data-toggle="tab"><i class="fa fa-cubes mx-2"></i>Notes Summary</a></li>
+          <li><a href="#ingRep" id="reps_tab" role="tab" data-toggle="tab"><i class="fa fa-exchange-alt mx-2"></i>Replacements</a></li>
+          <li><a href="#attachments" id="attachments_tab" role="tab" data-toggle="tab"><i class="fa fa-paperclip mx-2"></i>Attachments</a></li>
+          <li><a href="#revisions" id="revisions_tab" role="tab" data-toggle="tab"><i class="fa fa-clock-rotate-left mx-2"></i>Revisions</a></li>
+          <li><a href="#timeline" id="timeline_tab" role="tab" data-toggle="tab"><i class="fa fa-timeline mx-2"></i>History</a></li>
 
-          <li><a href="#formula_settings" id="formula_settings_tab" role="tab" data-toggle="tab"><i class="fa fa-cogs mr2"></i>Settings</a></li>
+          <li><a href="#formula_settings" id="formula_settings_tab" role="tab" data-toggle="tab"><i class="fa fa-cogs mx-2"></i>Settings</a></li>
         </ul>
                      
         <div class="tab-content">
@@ -261,7 +261,6 @@ function setProtected(status) {
 		dataType: 'json',
 		success: function (data) {
 			if ( data.success ) {
-				var msg = '<div class="alert alert-success alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a>' + data.success + '</div>';
 				fetch_formula();
 				if( data.success == 'Formula locked'){
 					$('#lock_status').html('<a class="fas fa-lock" href="javascript:setProtected(\'false\')">');
@@ -271,9 +270,8 @@ function setProtected(status) {
 					$('#add_ing').show();
 				}
 			} else {
-				var msg = '<div class="alert alert-danger alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a><strong>' + data.error + '</strong></div>';
+				$('#msgInfo').html('<div class="alert alert-danger alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a><strong>' + data.error + '</strong></div>');
 			}
-			$('#msgInfo').html(msg);
 		}
 	  });
 };